Use JavaScript classes instead of objects and arrays
