--- assets/js/add-book.js
@@ -1,8 +1,0 @@
-const addBook = (arr = [], title = '', author = '', id = '') => {
-  if (title && author) {
-    arr.push({ id, title, author });
-    return arr;
-  }
-  return false;
-};
-export default addBook;
