--- assets/js/data.js
@@ -1,19 +1,0 @@
-const initialData = [
-  {
-    id: '176543',
-    title: "Santa's Apprentice",
-    author: 'Nicolle Plackstone',
-  },
-  {
-    id: '223244556',
-    title: 'History Is Made at Night',
-    author: 'Eduard Ekell',
-  },
-  {
-    id: '3324',
-    title: 'Hercules',
-    author: 'Ofelia Lockless',
-  },
-];
-
-export default initialData;
