+++ assets/js/index.js
@@ -1,21 +1,14 @@
-import initialData from './data.js';
-import addBook from './add-book.js';
+/* eslint-disable max-classes-per-file */
 
+const form = document.querySelector('#form');
 const title = document.querySelector('#title');
 const author = document.querySelector('#author');
-const form = document.querySelector('.form');
-const ul = document.querySelector('.book-list');
-
-const saveToLocalStorage = (data) => {
-  const booksString = JSON.stringify(data);
-  localStorage.setItem('bookStoreData', booksString);
-  return true;
-};
 
 // generate unique id
 const generateId = (length = 10) => {
   let result = '';
-  const characters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
+  const characters =
+    'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
   const charactersLength = characters.length;
   for (let i = 0; i < length; i += 1) {
     result += characters.charAt(Math.floor(Math.random() * charactersLength));
@@ -23,57 +16,67 @@ const generateId = (length = 10) => {
   return result;
 };
 
-const savedBooks = () => {
-  const rawBooksData = localStorage.getItem('bookStoreData');
-  let books;
-  if (rawBooksData) {
-    books = JSON.parse(rawBooksData);
-  } else {
-    books = initialData;
-    saveToLocalStorage(books);
+class BookList {
+  constructor() {
+    this.data = [];
   }
-  return books;
-};
 
-const booksData = savedBooks();
+  addBook(book) {
+    this.data.push(book);
+    localStorage.setItem('bookStoreData', JSON.stringify(this.data));
+  }
 
-const getBooks = (myBooksArr) => {
-  const booksArr = myBooksArr.map((book) => {
-    const li = document.createElement('li');
-    const removeBtn = document.createElement('button');
-    removeBtn.innerText = 'Remove Book';
-    ul.append(li);
-    li.innerText = `${book.title} by ${book.author}`;
-    li.append(removeBtn);
-    removeBtn.setAttribute('id', book.id);
-    removeBtn.setAttribute('class', 'remove-button');
+  removeBook(id) {
+    const book = document.getElementById(id);
+    book.remove();
+    this.data = this.data.filter((book) => book.id !== id);
+    localStorage.setItem('bookStoreData', JSON.stringify(this.data));
+  }
+}
 
-    // remove element on click
-    removeBtn.addEventListener('click', (event) => {
-      for (let i = 0; i < myBooksArr.length; i += 1) {
-        if (myBooksArr[i].id === event.target.id) {
-          myBooksArr.splice(i, 1);
-          window.location.reload();
-          return saveToLocalStorage(myBooksArr);
-        }
-      }
-      return false;
-    });
+class Book {
+  constructor(title, author) {
+    this.title = title;
+    this.author = author;
+    this.id = generateId();
+  }
+}
 
-    // return created elements
-    return {
-      li,
-      removeBtn,
-    };
-  });
-  return booksArr;
-};
+const booksData = new BookList();
+
+function getBook() {
+  const book = new Book(title.value, author.value);
+  title.value = '';
+  author.value = '';
+  return book;
+}
+
+function displayBooks(book) {
+  const ul = document.querySelector('.book-list');
+  const li = document.createElement('li');
+  li.classList.add('list');
+  li.setAttribute('id', book.id);
+  li.innerText = `${book.title} by ${book.author}`;
+  const removeBtn = document.createElement('button');
+  removeBtn.innerText = 'Remove';
+  removeBtn.addEventListener('click', () => booksData.removeBook(book.id));
+  li.appendChild(removeBtn);
+  ul.appendChild(li);
+}
 
 form.addEventListener('submit', (e) => {
+  const newBook = getBook();
+  booksData.addBook(newBook);
+  displayBooks(newBook);
   e.preventDefault();
-  addBook(booksData, title.value, author.value, generateId());
-  window.location.reload();
-  return saveToLocalStorage(booksData);
 });
 
-getBooks(booksData);
+window.onload = () => {
+  booksData.data = JSON.parse(localStorage.getItem('bookStoreData' || '[]'));
+  if (booksData.data === null) {
+    booksData.data = [];
+    return;
+  }
+
+  booksData.data.forEach((book) => displayBooks(book));
+};
